components/auth-lib-dev/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auth-lib"
version = "0.1.0"
description = "Reusable FastAPI JWT auth/permissions library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt>=2.11.0",
    "fastapi>=0.100.0",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.28.0",
]

[[tool.uv.index]]
name = "nexus"
url = "https://nexus.dmi-msk.ru/repository/pypi-dev-group/simple"


[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::jwt.warnings.InsecureKeyLengthWarning"]

[tool.hatch.build.targets.wheel]
packages = ["auth_lib"]
